Moving stuff around and fixing missing jquery
